+++ logging.tf
@@ -2,17 +2,6 @@ resource "aws_s3_bucket" "bucket_logs" {
   bucket        = "${var.infrastructurename}-logs"
   tags          = var.tags
   force_destroy = true
-
-  #[S3.9] S3 bucket server access logging should be enabled
-
-}
-
-
-resource "aws_s3_bucket_logging" "logging" {
-  bucket = aws_s3_bucket.bucket_logs.id
-  #[S3.9] S3 bucket server access logging should be enabled
-  target_bucket = aws_s3_bucket.bucket_logs.id
-  target_prefix = "logs/bucket/${aws_s3_bucket.bucket_logs.id}"
 }
 
 resource "aws_s3_bucket_public_access_block" "buckets_logs_access" {
@@ -23,11 +12,9 @@ resource "aws_s3_bucket_public_access_block" "buckets_logs_access" {
   restrict_public_buckets = true
 }
 
-# [S3.5] S3 buckets should require requests to use Secure Socket Layer
-resource "aws_s3_bucket_policy" "buckets_logs_ssl" {
-
+resource "aws_s3_bucket_policy" "log_bucket_policy" {
   bucket = aws_s3_bucket.bucket_logs.bucket
-  policy = templatefile("${path.module}/templates/bucket_policy.json", { bucket = aws_s3_bucket.bucket_logs.bucket })
+  policy = jsonencode(local.log_bucket_policy)
 }
 
 resource "aws_s3_bucket_server_side_encryption_configuration" "bucket_logs_encryption" {
@@ -35,14 +22,58 @@ resource "aws_s3_bucket_server_side_encryption_configuration" "bucket_logs_encry
 
   rule {
     apply_server_side_encryption_by_default {
-      sse_algorithm = "aws:kms"
+      sse_algorithm     = "aws:kms"
+      kms_master_key_id = aws_kms_key.s3_logs_key.arn
     }
   }
 }
 
+resource "aws_kms_key" "s3_logs_key" {
+  description         = "KMS key for encrypting S3 access logs"
+  enable_key_rotation = true
+  tags                = var.tags
+  policy              = <<POLICY
+  {
+    "Version" : "2012-10-17",
+    "Id"      : "s3-logs-key-policy",
+    "Statement" : [
+      {
+        "Sid"    : "AllowRootAccountFullAccess",
+        "Effect" : "Allow",
+        "Principal" : {
+          "AWS" : "arn:aws:iam::${local.account_id}:root"
+        },
+        "Action"   : "kms:*",
+        "Resource" : "*"
+      },
+      {
+        "Sid"    : "AllowS3LoggingServiceToEncrypt",
+        "Effect" : "Allow",
+        "Principal" : {
+          "Service" : "logging.s3.amazonaws.com"
+        },
+        "Action" : [
+          "kms:Encrypt",
+          "kms:GenerateDataKey",
+          "kms:DescribeKey"
+        ],
+        "Resource" : "*",
+        "Condition" : {
+          "StringEquals" : {
+            "aws:SourceAccount" : "${local.account_id}"
+          }
+        }
+      }
+    ]
+  }
+  POLICY
+}
+
+
 resource "aws_kms_key" "kms_key_cloudwatch_log_group" {
   description         = "KMS key used to encrypt Kubernetes, VPC Flow, Amazon RDS for PostgreSQL and SSM Patch manager log groups within infrastructure ${var.infrastructurename}"
   enable_key_rotation = true
+  tags                = var.tags
   policy              = <<POLICY
 {
     "Version": "2012-10-17",