+++ modules/ivs_aws_instance/storage.tf
@@ -1,13 +1,29 @@
 resource "aws_s3_bucket" "data_bucket" {
-  bucket        = var.dataBucketName
+  count         = var.data_bucket.create ? 1 : 0
+  bucket        = var.data_bucket.name
   tags          = var.tags
   force_destroy = var.enable_deletion_protection ? false : true
 }
 
+resource "aws_s3_bucket_logging" "data_bucket_logging" {
+  count         = var.data_bucket.create ? 1 : 0
+  bucket        = aws_s3_bucket.data_bucket[0].id
+  target_bucket = var.log_bucket
+  target_prefix = "logs/bucket/${aws_s3_bucket.data_bucket[0].id}/"
+}
+
 resource "aws_s3_bucket" "rawdata_bucket" {
-  bucket        = var.rawDataBucketName
+  count         = var.raw_data_bucket.create ? 1 : 0
+  bucket        = var.raw_data_bucket.name
   tags          = var.tags
   force_destroy = var.enable_deletion_protection ? false : true
+}
+
+resource "aws_s3_bucket_logging" "rawdata_bucket_logging" {
+  count         = var.raw_data_bucket.create ? 1 : 0
+  bucket        = aws_s3_bucket.rawdata_bucket[0].id
+  target_bucket = var.log_bucket
+  target_prefix = "logs/bucket/${aws_s3_bucket.rawdata_bucket[0].id}/"
 }
 
 resource "aws_iam_role_policy" "eks_node_s3_access_policy" {
@@ -31,14 +47,9 @@ resource "aws_iam_role_policy" "eks_node_s3_access_policy" {
             ],
             "Effect": "Allow",
             "Resource": [
-                "${aws_s3_bucket.data_bucket.arn}",
-                "${aws_s3_bucket.rawdata_bucket.arn}"
-            ],
-            "Condition": {
-                "StringEquals": {
-                    "${local.goofys_user_agent_name}"
-                }
-            }
+                "${local.data_bucket_arn}",
+                "${local.raw_data_bucket_arn}"
+            ]
         },
         {
             "Action": [
@@ -49,14 +60,9 @@ resource "aws_iam_role_policy" "eks_node_s3_access_policy" {
             ],
             "Effect": "Allow",
             "Resource": [
-                "${aws_s3_bucket.data_bucket.arn}/*",
-                "${aws_s3_bucket.rawdata_bucket.arn}/*"
-            ],
-            "Condition": {
-                "StringEquals": {
-                    "${local.goofys_user_agent_name}"
-                }
-            }
+                "${local.data_bucket_arn}/*",
+                "${local.raw_data_bucket_arn}/*"
+            ]
         },
         {
             "Action": [
@@ -65,14 +71,9 @@ resource "aws_iam_role_policy" "eks_node_s3_access_policy" {
             ],
             "Effect": "Allow",
             "Resource": [
-                "${aws_s3_bucket.data_bucket.arn}/*",
-                "${aws_s3_bucket.rawdata_bucket.arn}/*"
-            ],
-            "Condition": {
-                "StringEquals": {
-                    "${local.goofys_user_agent_name}"
-                }
-            }
+                "${local.data_bucket_arn}/*",
+                "${local.raw_data_bucket_arn}/*"
+            ]
         }
     ]
 }
@@ -122,8 +123,8 @@ resource "aws_iam_role_policy" "s3_access" {
                 ],
                 "Effect": "Allow",
                 "Resource": [
-                    "${aws_s3_bucket.data_bucket.arn}",
-                    "${aws_s3_bucket.rawdata_bucket.arn}"
+                    "${local.data_bucket_arn}",
+                    "${local.raw_data_bucket_arn}"
                 ]
             },
             {
@@ -135,8 +136,8 @@ resource "aws_iam_role_policy" "s3_access" {
                 ],
                 "Effect": "Allow",
                 "Resource": [
-                    "${aws_s3_bucket.data_bucket.arn}/*",
-                    "${aws_s3_bucket.rawdata_bucket.arn}/*"
+                    "${local.data_bucket_arn}/*",
+                    "${local.raw_data_bucket_arn}/*"
                 ]
             },
             {
@@ -145,8 +146,8 @@ resource "aws_iam_role_policy" "s3_access" {
                 ],
                 "Effect": "Allow",
                 "Resource": [
-                    "${aws_s3_bucket.data_bucket.arn}/*",
-                    "${aws_s3_bucket.rawdata_bucket.arn}/*"
+                    "${local.data_bucket_arn}/*",
+                    "${local.raw_data_bucket_arn}/*"
                 ]
             }
         ]