+++ modules/simphera_aws_instance/minio-storage.tf
@@ -1,8 +1,7 @@
 resource "aws_iam_role" "minio_iam_role" {
   name        = "${local.instancename}-s3-role"
   description = "IAM role for the MinIO service account"
-  #depends_on  = [aws_iam_policy.minio_policy]
-  tags = var.tags
+  tags        = var.tags
   assume_role_policy = jsonencode({
     "Version" : "2012-10-17",
     "Statement" : [
@@ -37,8 +36,8 @@ resource "aws_iam_policy" "minio_policy" {
 }
 
 resource "aws_iam_role" "executor_role" {
-  name = "${var.name}-executoragentlinux"
-  #depends_on = [ aws_iam_policy.minio_policy ]
+  name        = "${var.name}-executoragentlinux"
+  description = "AWS IAM Role for the Kubernetes service account minio-irsa."
   # Terraform's "jsonencode" function converts a
   # Terraform expression result to valid JSON syntax.
   assume_role_policy = jsonencode({
@@ -94,7 +93,7 @@ resource "aws_s3_bucket_logging" "logging" {
   bucket = aws_s3_bucket.bucket.id
   #[S3.9] S3 bucket server access logging should be enabled
   target_bucket = var.log_bucket
-  target_prefix = "logs/bucket/${aws_s3_bucket.bucket.id}"
+  target_prefix = "logs/bucket/${aws_s3_bucket.bucket.id}/"
 }
 
 resource "aws_s3_bucket_versioning" "bucket_versioning" {