+++ modules/simphera_aws_instance/variables.tf
@@ -14,7 +14,7 @@ variable "postgresql_security_group_id" {
 }
 
 variable "tags" {
-  type        = map(any)
+  type        = map(string)
   description = "The tags to be added to all resources."
   default     = {}
 }
@@ -148,7 +148,7 @@ variable "log_bucket" {
 }
 
 variable "private_subnets" {
-  type        = list(any)
+  type        = list(string)
   description = "List of CIDRs for the private subnets."
   default     = ["10.1.0.0/22", "10.1.4.0/22", "10.1.8.0/22"]
 }